Obstacle when docking
A little feature where Liam will leave the BWF on its way home if it reaches an obstacle on the way.

If it hits the obstacle again within a given interval (OVERLOAD_INTERVAL) it will turn 90 degrees right (into the lawn) and start the mowing again. It will then attach to the BWF at another point on the lawn and sooner or later reach home.

This is a great feature for Islands. Just stick an obstacle on the way on the Island. Put the obstacle on a suitable point to send the mower in the optimal direction.