Turn at regular intervals
If the mower hasn't turned for TURN_INTERVAL, it may be because it has
gotten stuck somewhere. So back up a bit, turn and try to run again.
This is just a naive implementation, and may be improved on later.
Default time is set to 60 seconds for now.

Resolves #36.

Fun fact: This makes use of the global variable `time_at_turning` which
has been around - UNUSED - since the first commit to git, more than one
year ago...